+++ client/src/components/Login/Login.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Redirect, useHistory } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 import { connect } from "react-redux";
 import {
   Grid,
@@ -9,9 +9,9 @@ import {
   FormControl,
   TextField,
 } from "@material-ui/core";
-import { login } from "./store/utils/thunkCreators";
+import { login } from "../../store/utils/thunkCreators";
 import { makeStyles } from "@material-ui/core/styles";
-import bubble from "../src/components/SVG/bubble.svg";
+import { Theme, NavHeader } from "./index";
 
 export const useStyles = makeStyles((theme) => {
   return {
@@ -20,39 +20,11 @@ export const useStyles = makeStyles((theme) => {
       justifyContent: "space-between",
       margin: theme.spacing(2, 7, 2, 2),
     },
-    imageContainer: {
-      backgroundImage: `linear-gradient(to bottom, rgba(58, 141, 255, 0.85), rgba(134, 185, 255, 0.85)), url("../../images/bg-img.png")`,
-      padding: 0,
-      width: "100%",
-      backgroundSize: "cover",
-      backgroundRepeat: "no-repeat",
-      position: "relative",
-    },
-    img: {
-      width: "100%",
-    },
     mainContent: {
       display: "flex",
       flexDirection: "column",
       justifyContent: "flex-start",
       alignItems: "center",
-      marginTop: theme.spacing(5),
-    },
-    topBar: {
-      marginRight: theme.spacing(5),
-    },
-    sloganContainer: {
-      position: "absolute",
-      top: "35%",
-      left: "0",
-      width: "100%",
-    },
-    sloganText: {
-      fontSize: theme.typography.fontSize.title,
-      textAlign: "center",
-      color: "#ffffff",
-      margin: "auto",
-      width: "60%",
       marginTop: theme.spacing(5),
     },
     title: {
@@ -62,26 +34,6 @@ export const useStyles = makeStyles((theme) => {
     },
     loginBody: {
       marginTop: theme.spacing(9),
-    },
-    whiteButtonGrid: {
-      width: 154,
-      height: 54,
-      borderRadius: 5,
-      boxShadow: "0px 2px 10px rgba(74,106,149,0.2)",
-    },
-    whiteButton: {
-      width: 154,
-      height: 54,
-      fontSize: theme.typography.fontSize.medium,
-      textAlign: "center",
-    },
-    bubbleLogoContainer: {
-      display: "flex",
-      alignItems: "center",
-      width: "100%",
-    },
-    bubbleLogo: {
-      margin: "auto",
     },
     blueButtonGrid: {
       width: 160,
@@ -102,7 +54,6 @@ export const useStyles = makeStyles((theme) => {
 
 const Login = (props) => {
   const classes = useStyles();
-  const history = useHistory();
   const { user, login } = props;
 
   const handleLogin = async (event) => {
@@ -118,36 +69,12 @@ const Login = (props) => {
   }
   return (
     <Box className={classes.root}>
-      <Box className={classes.imageContainer}>
-        <Grid className={classes.sloganContainer}>
-          <Grid className={classes.bubbleLogoContainer}>
-            <img className={classes.bubbleLogo} src={bubble} alt="Logo" />
-          </Grid>
-          <Typography className={classes.sloganText}>
-            Converse with anyone with any language
-          </Typography>
-        </Grid>
-      </Box>
+      <Theme />
       <Grid container item className={classes.mainContent}>
-        <Grid container item justify="flex-end" alignItems="center">
-          <Typography
-            variant="body1"
-            color="secondary"
-            className={classes.topBar}
-          >
-            Don't have an account?
-          </Typography>
-          <Grid className={classes.whiteButtonGrid}>
-            <Button
-              size="large"
-              color="primary"
-              className={classes.whiteButton}
-              onClick={() => history.push("/register")}
-            >
-              Create account
-            </Button>
-          </Grid>
-        </Grid>
+        <NavHeader
+          title={"Don't have an account?"}
+          buttonText={"Create account"}
+        />
         <Grid
           direction="column"
           alignItems="center"
