+++ client/src/components/Login/index.js
@@ -1,0 +1,2 @@
+export { default as NavHeader } from "./TopNavHeader";
+export { default as Theme } from "./Theme";
