+++ client/src/routes.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Route, Switch, withRouter } from "react-router-dom";
 import { connect } from "react-redux";
 import { fetchUser } from "./store/utils/thunkCreators";
-import Signup from "./Signup.js";
-import Login from "./Login.js";
+import Signup from "./components/Login/Signup";
+import Login from "./components/Login/Login";
 import { Home, SnackbarError } from "./components";
 
 const Routes = (props) => {
